+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - [ ] JavaScript Intellisense enhancements
 - [ ] Add screenshot of JS intellisense to [readme](README.md#javascript)
 - [ ] Add missing JavaScript snippets
+- [ ] Implement telemetry
+- [ ] Implement logging
 
 Features that have a checkmark are complete and available for
 download in the